+++ private/buf/cmd/internal/internal.go
@@ -29,6 +29,7 @@ import (
 func GetModuleConfigForProtocPlugin(
 	ctx context.Context,
 	configOverride string,
+	module string,
 ) (bufconfig.ModuleConfig, error) {
 	bufYAMLFile, err := bufcli.GetBufYAMLFileForDirPathOrOverride(
 		ctx,
@@ -41,13 +42,19 @@ func GetModuleConfigForProtocPlugin(
 		}
 		return nil, err
 	}
+	if module == "" {
+		module = "."
+	}
 	for _, moduleConfig := range bufYAMLFile.ModuleConfigs() {
 		// If we have a v1beta1 or v1 buf.yaml, dirPath will be ".". Using the ModuleConfig from
 		// a v1beta1 or v1 buf.yaml file matches the pre-refactor behavior.
 		//
-		// If we have a v2 buf.yaml, we say that we need to have a module with dirPath of ".", otherwise
+		// If we have a v2 buf.yaml, users have to provide a module path or full name, otherwise
 		// we can't deduce what ModuleConfig to use.
-		if dirPath := moduleConfig.DirPath(); dirPath == "." {
+		if dirPath := moduleConfig.DirPath(); dirPath == module {
+			return moduleConfig, nil
+		}
+		if fullName := moduleConfig.ModuleFullName(); fullName != nil && fullName.String() == module {
 			return moduleConfig, nil
 		}
 	}