+++ windows/test.bash
@@ -2,9 +2,9 @@
 
 set -eo pipefail
 
-PROTOC_VERSION="27.0-rc1"
-PROTOC_GEN_GO_VERSION="v1.33.0"
-CONNECT_VERSION="v1.16.1"
+PROTOC_VERSION="27.0"
+PROTOC_GEN_GO_VERSION="v1.34.1"
+CONNECT_VERSION="v1.16.2"
 
 # Convert DOWNLOAD_CACHE from d:\path to /d/path
 DOWNLOAD_CACHE="$(echo "/${DOWNLOAD_CACHE}" | sed 's|\\|/|g' | sed 's/://')"