+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 ## [Unreleased]
 
-- No changes yet.
+- Allow user to override `--source-control-url` and `--create-default-label` when using
+  `--git-metadata` with `buf push`.
+- Fix `buf push --git-metadata` when local tags point to different objects than
+  the remote tags.
+- Fix issue where comment ignores were not respected for `PROTOVALIDATE` lint rule violations.
+
+## [v1.32.2] - 2024-05-28
+
+- Update `buf generate` to warn instead of error when proto3 optional is required but not
+  supported by a plugin.
+
+## [v1.32.1] - 2024-05-21
+
+- Fix archive and git inputs so that `--path` and `--exclude-path` paths are relative to
+  the `#subdir` rather than the root of the input. This fixes an unintended behavior change
+  that was introduced in `v1.32.0`.
+- Add `module` input for `protoc-gen-buf-lint` and `protoc-gen-buf-breaking` to allow
+  users to specify the module for `v2` configuration files.
 
 ## [v1.32.0] - 2024-05-16
 
@@ -1114,7 +1131,9 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.32.0...HEAD
+[Unreleased]: https://github.com/bufbuild/buf/compare/v1.32.2...HEAD
+[v1.32.2]: https://github.com/bufbuild/buf/compare/v1.32.1...v1.32.2
+[v1.32.1]: https://github.com/bufbuild/buf/compare/v1.32.0...v1.32.1
 [v1.32.0]: https://github.com/bufbuild/buf/compare/v1.32.0-beta.1...v1.32.0
 [v1.32.0-beta.1]: https://github.com/bufbuild/buf/compare/v1.31.0...v1.32.0-beta.1
 [v1.31.0]: https://github.com/bufbuild/buf/compare/v1.30.1...v1.31.0